pyproject.toml
[project]
name = "tradingagents"
version = "0.1.0"
description = "Multi‑Agents LLM Financial Trading Framework with OpenRouter integration"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  # Core dependencies
  "python-dotenv>=1.0.0",
  "typing-extensions>=4.14.0",

  # LangChain and OpenAI
  "langchain-openai>=0.3.23",
  "langchain-experimental>=0.3.4",
  "langgraph>=0.4.8",
  "openai>=1.14.2",

  # Data processing
  "pandas>=2.3.0",
  "yfinance>=0.2.63",
  "stockstats>=0.6.5",

  # Web scraping and APIs
  "requests>=2.32.4",
  "feedparser>=6.0.11",
  "praw>=7.8.1",

  # PDF generation
  "pdfkit>=1.0.0",
  "markdown2>=2.4.0",

  # Utilities
  "tqdm>=4.67.1",
  "pytz>=2025.2",
  "python-dateutil>=2.8.2",

  # Memory and embeddings
  "chromadb>=1.0.12",

  # Build tools
  "setuptools>=80.9.0",
]

[project.optional-dependencies]
data_sources = [
  "akshare>=1.16.98",        # Chinese financial data API
  "eodhd>=1.0.32",           # EOD Historical Data API
  "finnhub-python>=2.4.23",  # Finnhub market data SDK
  "tushare>=1.4.21",         # Chinese stock/data API
]
backtesting = [
  "backtrader>=1.9.78.123",  # Strategy backtesting framework
]
cli_tools = [
  "questionary>=2.1.0",      # Interactive CLI prompts
  "rich>=14.0.0",            # Rich terminal output
]
scraping = [
  "parsel>=1.10.0",          # HTML/XML scraping library
  "redis>=6.2.0",            # Caching and pub/sub
]
web_ui = [
  "chainlit>=2.5.5",         # Chat-based web UI framework
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
